完善大部分管理功能，规范化代码。

+++ src/main/java/st/southsea/blog/module/back/BArticleModule.java
@@ -1,4 +1,4 @@
-package st.southsea.blog.module;
+package st.southsea.blog.module.back;
 
 import org.nutz.dao.pager.Pager;
 import org.nutz.ioc.loader.annotation.Inject;
@@ -6,23 +6,18 @@ import org.nutz.ioc.loader.annotation.IocBean;
 import org.nutz.mvc.adaptor.JsonAdaptor;
 import org.nutz.mvc.annotation.*;
 import org.nutz.mvc.filter.CheckSession;
-import org.nutz.mvc.upload.TempFile;
-import org.nutz.mvc.upload.UploadAdaptor;
 import st.southsea.blog.base.Result;
 import st.southsea.blog.base.module.BaseModule;
 import st.southsea.blog.bean.Article;
+import st.southsea.blog.bean.User;
 import st.southsea.blog.service.ArticleService;
 import st.southsea.blog.service.LabelService;
 import st.southsea.blog.service.UserService;
 
 import javax.servlet.http.HttpSession;
-import java.io.IOException;
-import java.security.NoSuchAlgorithmException;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
-
-import static st.southsea.blog.module.UploadModule.SAVE;
 
 /**
  * @Author: South
@@ -31,7 +26,7 @@ import static st.southsea.blog.module.UploadModule.SAVE;
 @IocBean
 @At("/auth/article")
 @Filters(@By(type = CheckSession.class, args = {"admin", "/auth/login"}))
-public class ArticleModule extends BaseModule {
+public class BArticleModule extends BaseModule {
 
     @Inject
     private ArticleService articleService;
@@ -42,35 +37,25 @@ public class ArticleModule extends BaseModule {
     @Inject
     private LabelService labelService;
 
-    // 文章一览
+    // 主页
     @At("/")
     @Ok("beetl:/auth/article/index.html")
     public Object index(HttpSession session) {
         Map<String, Object> map = new HashMap<>();
-        map.put("user", userService.fetch((String) session.getAttribute("user")));
         map.put("labelList", labelService.query());
         return map;
     }
 
-    // 查询所有文章
+    // 表格
     @At("/list")
     @POST
     public Object list(Article article, int page, int limit) {
         Pager pager = new Pager(page, limit);
         List<Article> articleList = articleService.query(article, pager);
-        return layuiTable(articleList, articleService.count(article));
+        return layuiTable(articleList, articleService.count());
     }
 
-    // 写文章
-    @At("/edit")
-    @Ok("beetl:/auth/article/edit.html")
-    public Object edit() {
-        Map<String, Object> map = new HashMap<>();
-        map.put("labelList", labelService.query());
-        return map;
-    }
-
-    // 更新文章
+    // 编辑页面
     @At("/edit/?")
     @Ok("beetl:/auth/article/edit.html")
     public Object edit(long articleId) {
@@ -80,46 +65,62 @@ public class ArticleModule extends BaseModule {
         return map;
     }
 
-    // 图片接口
-    @At("/upload")
-    @POST
-    @Ok("raw:json")
-    @AdaptBy(type = UploadAdaptor.class, args = {"ioc:upload"})
-    public Object upload(HttpSession session, @Param("file") TempFile tempFile) throws IOException, NoSuchAlgorithmException {
-        String path = "../webapps/Blog_System/resources/article";
-        return SAVE(path, tempFile);
-    }
-
-    // 发表/更新文章
+    // 编辑文章
     @At
     @POST
     public Result edit(Article article) {
         if (articleService.fetch(article.getArticleId()) != null) {
             articleService.update(article);
-        } else {
-            articleService.insert(article);
+            ajaxSuccess("修改成功");
         }
-        return ajaxSuccess("文章发表成功");
+        return ajaxSuccess("修改失败");
     }
 
-    // 删除文章
+    // 编辑页面
+    @At("/add")
+    @Ok("beetl:/auth/article/add.html")
+    public Object add() {
+        Map<String, Object> map = new HashMap<>();
+        map.put("labelList", labelService.query());
+        return map;
+    }
+
+    // 编辑文章
     @At
     @POST
-    public Result delete(long articleId) {
-        if (articleService.fetch(articleId) == null) {
-            return ajaxError("该文章不存在");
+    public Result add(Article article) {
+        if (articleService.fetch(article.getArticleId()) == null) {
+            articleService.insert(article);
+            ajaxSuccess("发布成功");
         }
-        int i = articleService.delete(articleId);
-        return ajaxSuccess("删除成功");
+        return ajaxSuccess("发布失败");
     }
 
-    // 删除文章
+    // 单个删除
+    @At
+    @POST
+    public Result delete(long articleId, HttpSession session) {
+        User user = userService.fetch((String) session.getAttribute("user"));
+        // 判断是否为管理员
+        if (user.getPermissionId() == 1) {
+            articleService.delete(articleId);
+            return ajaxSuccess("删除成功");
+        }
+        return ajaxSuccess("删除失败");
+    }
+
+    // 批量删除
     @At
     @POST
     @AdaptBy(type = JsonAdaptor.class)
-    public Result deleteList(List<Article> articleList) {
-        int i = articleService.delete(articleList);
-        return ajaxSuccess("删除成功");
+    public Result delete(List<Article> articleList, HttpSession session) {
+        User user = userService.fetch((String) session.getAttribute("user"));
+        // 判断是否为管理员
+        if (user.getPermissionId() == 1) {
+            articleService.delete(articleList);
+            return ajaxSuccess("删除成功");
+        }
+        return ajaxSuccess("删除失败");
     }
 
 }
